pyproject.toml
[project]
name = "gyocas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.37.18",
    "python-dotenv>=1.0.1",
    "ruff>=0.11.2",
]

[tool.ruff.lint]
ignore = ["E203", "E226", "E241", "E242", "E266", "E402", "W391", "W505", "W605", "E701", "E731"]